# Speed up webpack build in development environment