--- a/package.json
+++ b/package.json
@@ -28,7 +28,8 @@
     "tslint-plugin-prettier": "^2.0.1",
     "typescript": "^3.2.4",
     "webpack": "^4.29.0",
-    "webpack-cli": "^3.2.1"
+    "webpack-cli": "^3.2.1",
+    "webpack-merge": "^4.2.1"
   },
   "dependencies": {
     "@google-cloud/datastore": "^3.0.1",