--- a/package.json
+++ b/package.json
@@ -17,6 +17,7 @@
     "@types/moment": "^2.13.0",
     "@types/moment-timezone": "^0.5.10",
     "@types/node": "^10.12.19",
+    "hard-source-webpack-plugin": "^0.13.1",
     "jest": "^23.6.0",
     "prettier": "^1.16.2",
     "prettier-tslint": "^0.4.2",