--- a/webpack.dev.js
+++ b/webpack.dev.js
@@ -1,6 +1,8 @@
 const merge = require('webpack-merge');
 const common = require('./webpack.common.js');
+const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
 
 module.exports = merge(common, {
-  mode: 'development'
+  mode: 'development',
+  plugins: [ new HardSourceWebpackPlugin() ]
 });