## 目的

localでの webpackのbuild時間を高速にしたい

https://webpack.js.org/guides/production/#simple-approach

- この資料を読むと、production環境は軽量にするべきと書かれている。
- しかしながら cloud functionsの用途であれば、production環境こそ devtoolを `source-map` として、エラーの内容の詳細を stack driverに送信したい
- そして簡単に動作確認をするために development環境のbuildは軽量にしたい

## 方法

- webpackのbuild方法について、時間が掛かってもsourcemapを残したいproduction buildと、時間を掛けずにサクサク buildしたい development buildの2つに分けた。
- localにbuildのcacheを作るwebpackのpluginを導入した。
  - https://github.com/mzgoddard/hard-source-webpack-plugin

### 参考資料
- webpackのplugin
  - https://github.com/mzgoddard/hard-source-webpack-plugin
- build optionを環境毎に変更する方法 
  - https://webpack.js.org/guides/production/#simple-approach
- devtoolの説明
  - https://webpack.js.org/configuration/devtool/

## 結果
### dev

```
$ time npm run build:dev
npm run build:dev  1.33s user 0.22s system 96% cpu 1.614 total
```

### prod (`"devtool" : "source-map"` の場合)

```
$ time npm run build
npm run build  5.60s user 0.44s system 133% cpu 4.529 total
```

### 備考 (`"devtool" : "none"` の場合)

```
$ time npm run build:dev
npm run build:dev  4.14s user 0.34s system 133% cpu 3.350 total
```

## その他

5秒くらいのbuild時間なら、`webpack-dev-server` で変更検知したらbuildくらいでも良いかもなぁ